Refactored StatViewCounts count from BigInteger to Integer

+++ src/main/java/com/jwt/services/StatsViewService.java
@@ -7,9 +7,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.jwt.enums.MessageTypes;
 import com.jwt.exceptions.MyMessageResponse;
 import com.jwt.models.*;
-import com.jwt.models.stats.*;
 import com.jwt.repositories.FixtureRepository;
-import com.jwt.repositories.StatNameRepository;
 import com.jwt.repositories.StatsViewRepository;
 import lombok.extern.java.Log;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -202,7 +200,7 @@ public class StatsViewService {
 
     private List<StatViewCounts> mapStatCountToStatViewCount( Map<String, Integer> averageStatCounts ) {
         return averageStatCounts.entrySet().stream()
-                .map(entry -> new StatViewCounts(entry.getKey(), BigInteger.valueOf(entry.getValue())))
+                .map(entry -> new StatViewCounts(entry.getKey(), entry.getValue()))
                 .sorted(Comparator.comparing(StatViewCounts::getCount).reversed())
                 .collect(Collectors.toList());
     }
